requirements.txt
boto3==1.18.1
botocore==1.21.1
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
python-ldap
requests
